Выносит фильтрацию в отдельный компонент
